Move functions to `persist`
